move genLink function to its own module
